requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
pyjwt>=2.8.0
slowapi>=0.1.9
httpx>=0.25.0
groq>=0.4.0
pyyaml>=6.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
